split into sub-protocols
